refactor - refactore domain
