+++ crates/domain/src/entities/blocklist.rs
