+++ crates/domain/src/entities/client.rs
