+++ crates/domain/src/entities/client_subnet.rs
