+++ crates/domain/src/entities/custom_service.rs
