+++ crates/domain/src/entities/group.rs
