+++ crates/domain/src/entities/managed_domain.rs
