+++ crates/domain/src/entities/mod.rs
@@ -1,0 +1,16 @@
+pub mod block_source;
+pub mod blocked_service;
+pub mod blocklist;
+pub mod blocklist_source;
+pub mod client;
+pub mod client_subnet;
+pub mod custom_service;
+pub mod group;
+pub mod managed_domain;
+pub mod query_log;
+pub mod regex_filter;
+pub mod safe_search;
+pub mod schedule;
+pub mod service_catalog;
+pub mod whitelist;
+pub mod whitelist_source;
