+++ crates/domain/src/entities/safe_search.rs
