+++ crates/domain/src/entities/schedule.rs
