+++ crates/domain/src/entities/service_catalog.rs
