+++ crates/domain/src/entities/whitelist.rs
