+++ crates/domain/src/errors/domain_error.rs
