+++ crates/domain/src/errors/mod.rs
@@ -1,0 +1,1 @@
+pub mod domain_error;
