+++ crates/domain/src/value_objects/dns_protocol.rs
