+++ crates/domain/src/value_objects/mod.rs
@@ -1,0 +1,5 @@
+pub mod dns_protocol;
+pub mod dns_query;
+pub mod dns_request;
+pub mod query_filters;
+pub mod validators;
